pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "celstate"
version = "0.1.0"
description = "AI-powered transparent asset generation CLI"
authors = [
    {name = "Emre", email = "emre@celstate.com"},
]
dependencies = [
    "google-genai>=1.0.0",
    "google-api-core>=2.0.0",
    "pillow>=10.0",
    "opencv-python-headless>=4.8",
    "numpy>=1.24",
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "python-multipart>=0.0.9",
    "python-dotenv>=1.0",
    "convex",
    "requests",
    "openai>=1.0",
    "typer>=0.12.0",
    "rich>=13.0",
]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

[project.scripts]
celstate = "celstate.cli.main:app"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/celstate"]

[tool.hatch.build.targets.sdist]
include = ["src/celstate"]

[tool.black]
line-length = 88
target-version = ['py312']

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
